basic structure for separate types
