+++ package.json
@@ -2,7 +2,8 @@
   "name": "cache-all",
   "version": "2.0.8",
   "description": "Fast, efficient cache engines for both express routes cache & native node (redis, in-memory & file cache)",
-  "main": "./index.js",
+  "main": "./memory.js",
+  "typings": "./memory.d.ts",
   "scripts": {
     "lint": "eslint ./src --fix",
     "test": "nyc mocha ./test/*.spec.js --exit",
