+++ redis.d.ts
@@ -1,0 +1,6 @@
+import Cacher from './inddex.d.ts';
+
+declare module 'cache-all/redis' {
+  const redis: Cacher;
+  export default Cacher;
+}
